Seperate operations dir
