+++ backend/crates/server/src/fhir_client/middleware/transaction.rs
@@ -9,18 +9,11 @@ use haste_fhir_client::{
     middleware::MiddlewareChain,
     request::{FHIRRequest, FHIRResponse},
 };
-use haste_fhir_model::r4::generated::{
-    resources::{Membership, Resource},
-    terminology::IssueType,
-};
+use haste_fhir_model::r4::generated::terminology::IssueType;
 use haste_fhir_operation_error::OperationOutcomeError;
 use haste_fhir_search::SearchEngine;
 use haste_fhir_terminology::FHIRTerminology;
-use haste_repository::{
-    Repository,
-    admin::ProjectAuthAdmin,
-    types::membership::{self as m, CreateMembership},
-};
+use haste_repository::Repository;
 use std::sync::Arc;
 use tracing::info;
